Refactor/clean up the entry point, to make future modifications easier.
The entry point, which was previously written mostly in Rust, is now split into two procedures. The first procedure is a pure-assembly procedure that sets up the stack pointer, and the second is a Rust function that performs higher-level setup. This fixes an issue with the existing setup: the existing _start function uses local variables (which may be placed on the stack) before it initializes the stack pointer.

The next step is to copy the .data section from flash into RAM and perform relocations on it. This may require modifying elf2tab to copy over relocation information.