Split the single command() system call into command0(), command1(), and command2().
The existing command() definition always sets both argument registers. For many uses of command(), this is unnecessary. Setting only the needed argument registers reduces the generated assembly size.

In its current form, this change is not backwards-compatible: users of the command() function will need to be updated to point to command0(), command1(), or command2() (with command2() being a drop-in replacement).