Re-enable support for synchronous main functions and fix `panic!()` in boards with no LED capsule.
This breaks the existing support for `async` `main()` functions. Instead, an `async_main!()` macro is provided to generate a synchronous `main()` that wraps the provided `async` `main()` function (which, unfortunately, must not be called `main()`).

I would prefer to be more consistent with Tokio and use a procedural macro attribute (as was originally done in #104), but I found that has some major drawbacks:
  1. If main() does not parse, the error message does not point to the correct code.
  2. It adds a requirement to update `syn` every time the toolchain is updated.
These drawbacks are elaborated upon in #111.

This solves #111 and #113.