Update flash.sh to better handle RISC-V targets
The current flash.sh script doesn't handle RISC-V targets which means that
even if you build with cargo, the actual tbf binaries don't get built.
Update the flash.sh script to handle this case.

As part of this work, also update the linker scripts for RISC-V targets.
These need to be updated to work the protected region argument in elf2tab