Get the application heap size from the environment
Currently, the application heap size is hard coded in entry_point/mod.rs.
If the app heap size needs to be adjusted, we have to manually change
the code and then updat elf2tab as well. Instead of relying on this
potentially troublesome process, allow taking the value from the environment
via APP_HEAP_SIZE. This is parsed at compile time.