core/entry_point: Start the heap at app_heap_break
Start the app's head size at the value of app_head_break passed by the
kernel.

This gives us more space to work with which allows us to create large
buffers (such as for passing to/from the HMAC).

It also means we do what the kernel expects us to do

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>