171 enable debug symbols in release build
Prior to this commit, the release profile was building with its default of leaving debug as false -- creating binaries that do not have DWARF debug sections. Without these sections (which are not loadable and therefore do not affect the size of the generated TBF), the resulting binary is effectively undebuggable:  it is optimized, heavily inlined (thereby eliding many symbols), and without any of the information necessary to connect generated code back to its source.  This commit brings libtock-rs inline with Tock itself (and many other Rust projects) which make debug explicit in the dev profile -- and set it in the release profile.