Set the artifacts workflow to only run when `master` is pushed to.
In the current setup, every time I push any branch to my fork I get an artifacts action run, which seems very wasteful. This change makes the actions workflow only run when master is modified. Note that the workflow will still run in forks, but only when the fork's master branch is updated.