Remove Allowed::replace and Allowed::take.
When I originally implemented Allowed, I made it work with non-Copy types. Allowed::get cannot with with non-Copy types, so I added Allowed::replace to allow client code to read a non-Copy type out of an Allowed buffer. However, during the PR review we changed Allowed to only contain Copy types. Because of this change, I don't anticipate Allowed::replace (and Allowed::take which is built on top of Allowed::replace) seeing much use. This PR removes Allowed::replace and Allowed::take to reduce the amount of `unsafe` in libtock_platferm.