console: Add support for println! and print!
Currently we have a cluncky way of printing by calling the following:

    writeln!(console, "My message")?;

Although this works, it means we need to keep track of the console
struct. This makes it difficult to support printing in third party
libraries. This also means that we don't follow the Rust convention of
supporting print! and println! macros.

This patch makes the console object a global static so that we can
support the print! and println! macros. All current users are converted
to use this new method.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>