Configure Bors to use GitHub Actions rather than Travis checks.
The GitHub Actions work better than Travis, and are a superset of the Travis checks. Travis is currently broken and I'd rather replace it with Actions than fix it. This PR should make Bors listen to GitHub actions rather than Travis.

If this works, then I'll remove the Travis config entirely in a future PR.