Remove the Travis config.
This finishes the move from Travis for CI to GitHub Actions. Everything Travis did is done by GitHub Actions as well, and Bors has been configured to use Actions rather than Travis to gate PR merges. The Travis CI is currently broken, this removes that CI config so we don't get failed status checks on PRs in the GitHub UI.