tools/flash.sh: Automatically generate the permissions
Update the `flash.sh` script to automatically generate the app
permissions. This requires the elf2tab PR: https://github.com/tock/elf2tab/pull/28
to accept the --permissions argument.

This script will parse the output to determine what capsules are being
used in the app. It then allows all of the commands in libtock-rs that
can be used by those capsules.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>