Overhaul libtock_platform's syscalls trait, making the following changes:
1. The trait is now design for the (work-in-progress) Tock 2.0 syscalls rather
   than the Tock 1.0 syscalls.
2. The higher-level API that I previously intended to implement in the Platform
   type is now expressed in the form of provided functions in Syscalls.

So far, I have only implemented yield. I'm sending this for review now because it is already getting large, and I'd prefer to split it up into several PRs rather than send a single huge PR.

There is still significant design work to be done in terms of error handling for subscribe, command, allows, and memop.