Trim down the Mac OS CI to just a `cargo build` on libtock_runtime.
The Mac runner took over 1 hour 15 minutes to build QEMU (and it didn't even succeed), so I don't think running `make setup` on Mac OS is reasonable. Shrink its test to `cargo build -p libtock_runtime`, which should be relatively simple and test the most critical items.

In the future, we'll probably want to point it at a more substantial crate, but libtock_runtime is our best option for now.