Move `libtock_platform` and `libtock_runtime` out of `core/`.
`libtock_platform` was originally built in `core/` because I was intending to replace `libtock_core`'s components one-by-one with replacements built on `libtock_platform`. However, this plan changed when we decided to build the new `libtock-rs` directly against Tock 2.0 rather than Tock 1.0, as that made `libtock_core` and `libtock_platform` fundamentally incompatible.

Instead, these crates should be in the root of the repository. When we remove Tock 1.0 support from `libtock-rs`, we will remove the `core/` directory entirely.

To make this change, I did:

1. `git mv core/platform .`
2. `git mv core/runtime .`
3. Remove `core/` from the crate paths in `./Cargo.toml`
4. Remove the `libtock_platform` dependency from `core/Cargo.toml`