Add the `libtock2` crate and clean up `make test` to better match the new crate organization.
`libtock2` is a "one stop shop" for using `libtock-rs`. It is intended to be easy to use and will provide everything needed to create a `libtock-rs` process binary. However, in order to be easy to use, it will not support unit testing.

I replaced the `libtock_runtime` build in `make test` with a build of `libtock2`, as that will eventually depend on most of the Tock 2.0 crates. I also cleaned up the `clippy` invocations a bit, and made them check `libtock2`.