Add the `syscalls_tests` crate.
Although `Syscalls` is implemented in `libtock_platform`, its unit tests cannot live in `libtock_platform`, due to a confusing "multiple copies of crate" issue that I explained in the comments. Instead, I'm putting its tests into a new crate.

This crate is currently called `syscalls_tests` because it contains tests for the `Syscalls` implementation. I'm happy to take other name suggestions -- I'm not super happy with this one, because it's a bit tricky to remember to add both 's'-es to the name. I don't want to do `platform_tests` because that would break tab completion for `platform/` and I think some `libtock_platform` tests belong in `libtock_platform`.