[RFC] Allow PR authors to merge their own PR.
This change only applies to PRs authored by `libtock-rs` code owners.

To date, we have asked PR reviewers to merge PRs. In practice, this means the PR author writes something like "This PR is ready to merge" when it is ready to merge, and the PR reviewers respond with `bors r+`. I do not believe we have ever had a situation where a PR reviewer has given any other response. This seems like an unnecessary back-and-forth.