Add safe constructors for `CommandReturn` to `libtock_unittest`.
`libtock_platform` does not provide safe constructors for `CommandReturn`, as in a process binary it is supposed to only be created by invocations of the Command system call. However, fake drivers for unit tests (as well as many unit tests) have a need to create their own `CommandReturn` values. To avoid a proliferation of `unsafe`, I added safe constructors for `CommandReturn` to `libtock_unittest`.

For testing purposes, I added unconditional `Debug` derives to some of the `libtock_platform` types. These are meant only for use in host-based tests: embedded code should use a lighter-weight solution (we're currently trying out `ufmt` for this).