Move `fake::Kernel`'s data into a new type, `KernelData`.
Prior to this change, `fake::Kernel` had two conflicting responsibilities:

1. It is the public interface to the fake kernel, used by unit tests.
2. It stores the fake kernel's data.

This initially appeared to be the correct design, but it interacts poorly with system calls. In particular, system calls are highly reentrant (a Yield call can invoke an upcall that can call other system calls), and can call other methods that use the fake kernel (e.g. to queue an upcall). This reentrancy can result in many `Rc<fake::Kernel>`s existing in separate stack frames, which gets confusing.

To make writing the reentrant code easier, this change pulls the data out into `KernelData`. `KernelData` is stored in a thread local variable that replaces the thread local variable that previously held a `Weak<fake::Kernel>`. `KernelData` is data-only by design (i.e. it lacks behavior), which makes it easier to write code that is reentrancy-safe.

Note that this change drops some of the `fake::Kernel` leak checking (namely, the check that there isn't a `fake::Kernel` present when thread-local storage is deallocated). That check is less necessary now that we no longer use `Rc` to own `fake::Kernel`, and is also quite a bit more verbose to implement.