Add an upcall module to `libtock_unittest`.
This change adds an upcall queue to the kernel data. I plan to implement the following in separate PRs:

1. Upcall invocation in Yield
2. Subscribe

I added a TODO to write a test for fake::Kernel::add_driver. There are several refactorings I want to do to it:

1. Make it take &mut self (along with most of the other fake::Kernel methods)
2. Make it take the driver's Rc by value.

If I write the test now, I'll have to change it significantly when I refactor add_driver, so I left it out.