Fix the aliasing invariants on Allow buffers.
TRD 104 says a process *shouldn't* access a buffer that has been shared with the kernel, but it is not *prohibited* from doing so. However, it does require the kernel to permit userspace library implementations that prohibit processes from accessing shared buffers. `libtock-rs` makes use of this prohibition: its Allow APIs are designed to work with `&mut [u8]` references and immutable references, rather than `&[Cell<u8>]` references.

The prohibition materializes as a safety invariant on `RawSyscalls`: its Allow implementation is more strict than TRD 104 requires. This commit adds that invariant to `RawSyscalls`.

Additionally, the "may be soundly converted to a slice" part of the `RoAllowBuffer` and `RwAllowBuffer` invariants was intended to imply that aliasing invariants are present. Instead of implying it, I added sentences to state it outright.