Add a document describing a safe API for Tock 2.0's Subscribe syscall.
Creating a safe, sound API for Subscribe that supports callbacks allocated on
the stack is nontrivial, because it is difficult to guarantee the subscription
is cleaned up before the callback is deallocated. This design, which is a
continuation of https://github.com/tock/libtock-rs/issues/338, is an attempt to
create a safe API for Subscribe.

It is inspired by `core::pin`, but for efficiency it uses a distinct handle type
(`SubscribeHandle`) that is zero-sized.

I intend to have this PR reviewed by reviewers who are unfamiliar with Tock, and
I wrote this document accordingly.

[Rendered](https://github.com/jrvanwhy/libtock-rs/blob/subscribe-api-design/doc/SubscribeApiDesign.md)