Remove the `ID` generic from `Callback`.
The advantage of having the `ID` parameter is it allows for a single type (e.g. a driver) to accept multiple callbacks. However, as was pointed out in the review comics, this causes `Callback::callback` (as well as the `extern "C"` shim that calls it) to be monomorphized for each subscribe ID, which adds code size.

Ti50 prefers a design where there are very few distinct Subscribe callbacks, each of which writes a value into a `Cell<Option<[u32; 3]>>` (or similar) on the stack. That design may result in a smaller code size, but increases RAM usage.

Removing `ID` removes the ability to accept multiple callbacks, but it also allows for Ti50 to achieve the tradeoffs they want to achieve. It's not clear to me whether there will be a single type that wants to listen to multiple callbacks, but if we need to we can add a disambiguator in the following manner without causing excessive monomorphization:

```rust
struct DefaultCallback;

// The Disambiguator type should be DefaultCallback *unless* this type
// implements multiple callbacks, in which case each callback would have
// a different Disambiguator type.
trait Callback<Disambiguator = DefaultCallback> {
    fn callback(&self, args: [u32; 3]);
}
```