Add `syscall_scope`, a scope guard for system calls.
`syscall_scope` creates a scope in which Allow and Subscribe system calls can be soundly executed, by guaranteeing that the buffers and upcalls are revoked before their lifetime ends. The exact system calls (type and ID) are inferred from the type of the closure passed to `syscall_scope`, using the `ShareList` trait.

`syscall_scope` is the `run` function described at https://github.com/tock/libtock-rs/issues/341, and `ShareList` is the `SyscallList` trait.