Update the Rust toolchain to 2021-05-27.
`cargo miri test` now appears to run doc tests, regardless of whether or not they are marked `#[cfg(not(miri))]`. That caused a test failure when `exit_test` was run. To fix this, I added some manual config annotations into the `exit_test` doc test to make it a no-op under Miri.

Changes in lints led to the following tweaks:

1. A number of `assert_eq!(..., <bool>)` checks were replaced with `assert!(...)`.
2. An unused field was removed from the `ble_scanning` example.
3. A `std::mem::replace` call was migrated to `std::mem::take`.