Refactor `syscall_scope` into a new `share` module with a different layout.
This refactoring is motivated by https://github.com/tock/tock/pull/2906, which removes the need for passing around references to `ShareList`s. Instead, this change introduces a `Handle` type, which functions like a `&ShareList` but which is zero-sized.

I noticed that a core concept of the various types/methods is that of sharing data with the kernel, so I renamed `syscall_scope` into `share_scope`. I ended up with a number of names that started with `share` (`share_scope`, `ShareList`, and `ShareHandle`), so I made `share` its own module.

This will make `libtock-rs` depend on https://github.com/tock/tock/pull/2906, so it should only be merged if https://github.com/tock/tock/pull/2906 is merged.