ufmt: add support for hex printing and width specifiers
This change adds ufmt support for hex printing of numbers and
for fixed-width printing up to widths of 18 using 0s or spaces
as padding. 0 padding is only supported for numbers. Custom fill
characters and specified alignment is not supported. This change
does decrease the size savings of ufmt relative to core::fmt,
but brings them much closer to feature parity.