Add a runner for libtock2 examples.
This replaces tools/flash.sh. In the future, it will have functionality to process test output, and will therefore replace test_runner as well.

It can currently deploy to Hifive1 on QEMU, and has untested logic to deploy via `tockloader`.

It is no longer possible to use `cargo run` to deploy a Tock 1.0 libtock-rs process binary, as that will now use the Tock 2.0 runner.

Additional changes:
1. added a `setup-qemu-2` build target to build the tock2 submodule's QEMU module. I intend to rename this to `setup-qemu` after the Tock 1.0 targets and submodules are removed.
2. I updated the tock2 submodule to a newer kernel. Because the new libtock-rs crates depend on post-2.0 Allow API guarantees, I updated to latest master.