libtock2: Enable overriding the toolchain for the assembler
After some testing and upstream support, a variant of this will be
upstreamed and this CL will be replaced by that. This uses the standard
Make flags AS, ASFLAGS, and AR. STRIP is non-standard but was confirmed
as fine to use by jrvanwhy.

This also fixes a particularly weird issue with using llvm-mc currently:
it cannot handle a comment on the same line as a `mv` instruction,
unless it ends in `;`. Sounds like a tokenizer bug.

BUG=b:210846576
TEST=make flash
SOURCE=CHROME-ONLY

Change-Id: I1a4c38cf918e43e4f3de45a0fb1851466f540c93