Add padding to `.start` so that `.text` is aligned to a multiple of 4 bytes.
I found that elf2tab was leaving out the padding between `.start` and `.text`. I *think* this is an elf2tab bug, but I'm not sure and regardless I want to be able to run libtock2 binaries without waiting for another elf2tab release. Therefore, I add the padding into the `.start` section directly via the linker script, which results in a correct TBF file.