Replace `_number` with `_num` for consistency.
As noticed in #373, this repository uses a mix of `_number` and `_num` for allow/command/driver/exit/subscribe numbers. We decided to use `_num` everywhere for consistency. This PR replaces the `_number` suffixes with `_num`.

There are still many uses of `_id` in the code from before TRD104 standardized on using "number" for those values. This PR is already getting large, so I'll leave that for a separate PR.