Clear the Thumb bit on `.start`'s address.
This is an alternative fix for the issue that https://github.com/tock/libtock-rs/pull/392 fixes. The underlying issue is that `rust-lld` (based on LLVM's LLD) and GCC's `ld` handle the Thumb bit differently. `rust-lld` sets the Thumb bit for symbol addresses, and `ld` does not, resulting in different `rt_header` values.

This PR resolves that issue by clearing the Thumb bit in the linker script when setting `rt_header.start`, and modifying the ARM startup assembly to no longer expect the Thumb bit to be set. Now, `libtock_runtime` should correctly start with either linker. This *should* be a no-op on RISC-V (where `.start` is always aligned to a power of 4 address), and was tested by hand on RISC-V using QEMU.