Update the Tock submodule to the latest dev version.
I was having trouble getting QEMU to build with the Tock version that this submodule had pinned. Because `libtock-rs` is designed for the upcoming Tock 2.1 (it depends on the new Allow API guarantees), it makes sense to update this module to the latest `master` branch commit. With the new submodule commit, I can now build QEMU.