runtime: set heap start and stack top at app startup
Add memop calls during application start up to set the heap starting
address and stack top address with the kernel. This does not take that
much extra time, but can be very valuable to the kernel if it is tracking
process debug information. These calls were also in 1.x version of
libtock-rs, so it seems appropriate to replicate them here as well.

Change-Id: I6e9b70922c98e37079730eec0c791a18db78bb43