Tell the kernel where the stack and heap begin in `rust_start`.
This is an alternative implementation of #406 which keeps all the changes in Rust.

**Pros relative to #406:** This does not add any more assembly code. It also allows the debug calls to be removed entirely via a feature flag.

**Cons relative to #406:** Adds 8 bytes to the code size to compute the address of `rt_header` (same size on ARM and RISC-V).

There is one more way to implement the debug calls, which is to add a single `mv` instruction to each entry point to move `rt_header`'s address into `a0` so that `rust_start` does not need to compute its address. That saves a couple bytes in the common case, but that instruction is not removable and it more tightly couples the entry point assembly with `rust_start`.

Last detail: I updated `start_prototype` to work with the new `RawSyscalls` trait, and ran `rustfmt` on it.