Use #[repr(u32)] for ErrorCode
This makes much more efficient use of the current default Rust ABI,
on RISC-V, which does not emit the `zeroext` attribute for a `u16` param.
This results in most code using it needing to zero out the top part of
the register, wasting 2 instructions (8 bytes) for virtually every use.

This change saves 3K of code size in Ti50.

Next step: investigate if changing the Rust ABI to emit `zeroext` for
non-word sized params generates more efficient code on RISC-V, as
truncation is free on x86/ARM and not on RISC-V.

Change-Id: I70543f7431aab73839d9524bdf0ff02c18d91c79