Remove our use of the `stack-sizes` tool.
As of version `0.5.0`, `stack-sizes` is [no longer a binary crate](https://github.com/japaric/stack-sizes/pull/14). This breaks our CI, which tries to install the latest version of the `stack-sizes` binary.

This PR removes all references to the tool from `libtock-rs`, so that CI can continue working.

`stack-sizes` appears to have been replaced by [`cargo-call-stack`](https://github.com/japaric/cargo-call-stack), which we may add support for in the future.