Increase the TBF header size from 0x48 to 0x60.
As of https://github.com/tock/elf2tab/pull/43, `elf2tab` generates a program header by default, increasing the size of the TBF headers to 96 bytes (`0x60`). This PR increases the space that `libtock-rs` gives the TBF headers by default to 96 bytes so it will work with the next `elf2tab` version.