Remove bors-ng and switch to GitHub actions merge-queue
This removes the bors configuration and switches to the GitHub actions
merge queue merge bot. It replicates the required status checks for
bors to pass, and adds the appropriate GitHub merge group status
checks. This PR needs to be followed up with the appropriate
configuration applied in the repository settings.