Define a MSRV, use that toolchain by default.
This PR defines a Minimum Supported Rust Version. It specifies that toolchain in `rust-toolchain`, so `rustup` will use that toolchain by default. The `test-stable` make action is removed, as it is now redundant.

We still need the nightly toolchain for Miri. Instead of specifying the nightly toolchain version via the command line, I created a new `nightly/rust-toolchain.toml` file that specifies the toolchain. The benefit is this makes `rustup` automatically install the nightly toolchain *including Miri*. This should stop the toil we currently have where every time the Rust toolchain is updated, everyone using `make test` has to manually install new toolchains and Miri. The downside is it's kinda ugly.

I set the MSRV to 1.70 so this doesn't conflict with #498.

I will send another PR that adds `rust-version.workspace = true` to all the `Cargo.toml` files in this repository; I didn't want to clutter this PR with those changes.