Update the `tock/` submodule to latest `master`.
Tock 2.1.1 uses a Rust toolchain that is too old to understand `libtock-rs`' current `Cargo.toml`. This causes `cargo` commands in the Tock submodule to fail. This PR updates the Tock submodule to latest `master`, giving it a recent toolchain. This fixes `make clean` and likely other build commands.

This also gives access to the license checker, which we may want to use in `libtock-rs`.