shell.nix: Add nix shell
Specifies a Nix shell that creates a deterministic development
environment. This is based on the Tock Nix shell but is modified to
combine the required dependencies between the dual stable and nightly
toolchains. This `shell.nix` is meant for developing, so it uses the nightly
Rust version in order to be able to use Miri.

Co-authored-by: Anthony Tarbinian <atar137h@gmail.com>